Reform our approach to the opposite category
I want to add preservation of isomorphisms as a property of functors,
however the [C.Op] module is used as a source category for contravariant
functors, and it does not contain the full [CategoryTheory] where [iso]
is defined. Under the new approach, the opposite category has to be added or
requested explicitly, but it is a full [Category] instance.